+++ models/employee.js
@@ -45,7 +45,7 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: false,
       },
       phone: {
-        type: DataTypes.INTEGER(15),
+        type: DataTypes.INTEGER(20),
         allowNull: false,
       },
       address: {
